Adding Image Upload-Download / PDF Making And Downloading

+++ controllers/admin.js
@@ -1,12 +1,13 @@
-// Import dependencies
+const mongoose = require('mongoose');
+const fileHelper = require('./file');
+
 const {
   validationResult
 } = require('express-validator');
+
 const Product = require('../models/product');
 
-// Handle GET request to add a new product
 exports.getAddProduct = (req, res, next) => {
-  // Render the "edit-product" template with initial data
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
@@ -17,18 +18,29 @@ exports.getAddProduct = (req, res, next) => {
   });
 };
 
-// Handle POST request to add a new product
 exports.postAddProduct = (req, res, next) => {
-  // Extract form data from the request
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasErrors: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'No Image Attached Or Attached file is not an image.',
+      validationErrors: []
+    });
+  }
 
-  // Validate the form data using "express-validator"
+  console.log(image)
   const errors = validationResult(req);
-
-  // If there are validation errors, render the "edit-product" template with error messages and previously entered data
   if (!errors.isEmpty()) {
     console.log(errors.array());
     return res.status(422).render('admin/edit-product', {
@@ -38,7 +50,6 @@ exports.postAddProduct = (req, res, next) => {
       hasErrors: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -47,7 +58,8 @@ exports.postAddProduct = (req, res, next) => {
     });
   }
 
-  // If there are no validation errors, create a new product using the "Product" model and save it to the database
+  const imageUrl = image.path;
+
   const product = new Product({
     title: title,
     price: price,
@@ -62,22 +74,18 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      // If there is an error during the process, pass the error to the error handler middleware
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(err);
+      console.log(err);
+      return next(error);
     });
 };
 
-// Handle GET request to edit an existing product
 exports.getEditProduct = (req, res, next) => {
-  // Check if the "edit" parameter is present in the query string
   const editMode = req.query.edit;
   if (!editMode) {
-    // If not, redirect the user to the home page
     return res.redirect('/');
   }
-  // Find the product with the given ID and render the "edit-product" template with the product data
   const prodId = req.params.productId;
   Product.findById(prodId)
     .then(product => {
@@ -95,26 +103,21 @@ exports.getEditProduct = (req, res, next) => {
       });
     })
     .catch(err => {
-      // If there is an error during the process, pass the error to the error handler middleware
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(err);
+      console.log(err);
+      return next(error);
     });
 };
 
-// Handle POST request to update an existing product
 exports.postEditProduct = (req, res, next) => {
-  // Extract form data from the request
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
-
-  // Validate the form data using "express-validator"
   const errors = validationResult(req);
 
-  // If there are validation errors, render the "edit-product" template with error messages and previously entered data
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Edit Product',
@@ -123,7 +126,6 @@ exports.postEditProduct = (req, res, next) => {
       hasErrors: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -133,7 +135,6 @@ exports.postEditProduct = (req, res, next) => {
     });
   }
 
-  // If there are no validation errors, find the product with the given ID, update its fields with the new data, and save it to the database
   Product.findById(prodId)
     .then(product => {
       if (product.userId.toString() !== req.user._id.toString()) {
@@ -142,23 +143,25 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path;
+      }
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
       });
     })
     .catch(err => {
-      // If there is an error during the process, pass the error to the error handler middleware
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(err);
+      console.log(err);
+      return next(error);
     });
 };
 
-// Handle GET request to get all products created by the currently logged in user
 exports.getProducts = (req, res, next) => {
-  // Find all products created by the currently logged in user and render them using a template engine
   Product.find({
       userId: req.user._id
     })
@@ -171,29 +174,34 @@ exports.getProducts = (req, res, next) => {
       });
     })
     .catch(err => {
-      // If there is an error during the process, pass the error to the error handler middleware
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(err);
+      console.log(err);
+      return next(error);
     });
 };
 
-// Handle POST request to delete a product with the given ID created by the currently logged in user
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  // Find the product with the given ID and delete it from the database
-  Product.deleteOne({
-      _id: prodId,
-      userId: req.user._id
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not Found!'))
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({
+        _id: prodId,
+        userId: req.user._id
+      })
     })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
     .catch(err => {
-      // If there is an error during the process, pass the error to the error handler middleware
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(err);
+      console.log(err);
+      return next(error);
     });
 };
